part of patch by David Evans to fix build.  Should be fixed differently in trunk in my opinion, but here is ok ;) Last part of the patch!
Fixed bugs:
  - https://launchpad.net/bugs/1245722

(bzr r9983)

+++ src/widgets/eek-preview.cpp
@@ -235,7 +235,7 @@ gboolean eek_preview_expose_event( GtkWidget* widget, GdkEventExpose* event )
 
         GdkGC *gc = gdk_gc_new( widget->window );
         EekPreview* preview = EEK_PREVIEW(widget);
-        GdkColor fg = {0, preview->_r, preview->_g, preview->_b};
+        GdkColor fg = {0, static_cast<guint16>(preview->_r), static_cast<guint16>(preview->_g), static_cast<guint16>(preview->_b)};
 
         gdk_colormap_alloc_color( gdk_colormap_get_system(), &fg, FALSE, TRUE );
